move pointer_uses_may_touch onto memory alias analysis
